Add user associations and other modifications to various models
The commit introduces user associations to the Flights, FlightFiles, and LogFiles models and also assigns the first user as the default owner for all instances of these models. A constraint on uuid and upload actions to ActiveGraph are added to the migration files. Additionally, a trajectory_id column to Flights and the creation of ActiveStorage tables have been implemented. It also includes the creation of devise users and adding a description field to Flights model.