Add user registration and authentication features
This commit adds Devise to handle user authentication along with necessary views, controllers, and routes. The User model has been created to store user data and relationships with other models. As part of this, a helper module 'CypherQueries' is introduced to define associated queries. Changes are made in the ApplicationController to handle user redirection based on their login status. Required gems are added to application.rb along with configurations in devise.rb. Also, devise.en.yml is added for storing error and success messages relevant to user registration and authentication. This feature enhances the application's security and enables user-tracking.